update Java version

+++ src/test/java/com/upokecenter/test/CBORTest.java
@@ -2324,7 +2324,7 @@ cmpCobj.toString() +
         }
       }
       for (int i = 2; i <= 5; ++i) {
-        int eb = (byte)(0x20 * i);
+        int eb = 0x20 * i;
         bytes = new byte[] { (byte)eb };
         try {
           CBORObject.DecodeFromBytes(bytes, options);
